% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ContrastSplines.R
\name{ContrastSplines}
\alias{ContrastSplines}
\title{ContrastSplines: Returns Spline matrix for full rank design matrix.}
\usage{
ContrastSplines(SB)
}
\arguments{
\item{SB}{[\code{numeric}] raw (Spline) base. Matrix whose rows all sum to one}
}
\description{
The function \code{ContrastSplines}  returns a matrix with spline base 
similar to "contr.sum", i.e. spline regression parameters sum to zewro. I.e. 
the sums of rows in the input does not all sum to one.
}
\details{
...
}
\examples{
\dontrun{

# Might come

}


}
\author{
Lars Erik Gangsei\cr
lars.erik.gangsei@vetinst.no\cr
+47 950 61 231
}
