% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DLM_Get_GF.R
\name{DLM_Get_GF}
\alias{DLM_Get_GF}
\title{DLM_Get_GF: Construct System matrix (G) and Design matrix (F) DGLM for Dynamic Linear Model (DLM).}
\usage{
DLM_Get_GF(
  q = 1,
  trend = FALSE,
  harmonic = FALSE,
  period = 365,
  freq = 1,
  n_waves = 1
)
}
\arguments{
\item{q}{[\code{numeric}]. The number/ length of the multivariate response.}

\item{trend}{[\code{logical}]. Set till "TRUE" if (one common) trend is to be included.}

\item{harmonic}{[\code{logical}]. Set till "TRUE" if (one common) seasonal effect is yto be included.}

\item{period}{[\code{numeric}]. Time period for the harmonic wave, I.e. the number of 
time units in one full cycle.}

\item{freq}{[\code{numeric}]. The number of peaks within each time period.}

\item{n_waves}{[\code{numeric}]. The number of harmonic waves.}
}
\value{
A System matrix G with dimension [q + trend == TRUE + 2 x n_waves] x [q + trend == TRUE + 2 x n_waves]
and a design matrix F with dimension [q + trend == TRUE + 2 x n_waves] x q
}
\description{
The function \code{DLM_Get_G} constructs the G and F matrices. It is (by now) 
possible to include a multivariate response with common trend and common seasonality 
modeled as harmonic waves.
}
\details{
...
}
\examples{
\dontrun{

}


}
\author{
Lars Erik Gangsei\cr
lars.erik.gangsei@vetinst.no\cr
+47 950 61 231
}
