% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DLM_Get_W.R
\name{DLM_Get_W}
\alias{DLM_Get_W}
\title{DLM_Get_W: Construct System matrix Covariance for Dynamic Linear Model (DLM).}
\usage{
DLM_Get_W(
  q = 1,
  matern = FALSE,
  dist = NULL,
  kappa = 10,
  nu = 1/5,
  sigma_d2 = 1,
  sigma_t2 = 1,
  trend = FALSE,
  harmonic = FALSE,
  n_waves = 1,
  regions = FALSE
)
}
\arguments{
\item{q}{[\code{numeric}]. The number/ length of the multivariate response.}

\item{matern}{[\code{logical}]. If covariance between multivariate should modelled as matern Covariances}

\item{dist}{[\code{matrix}]. Either a distance matrix [q x q] between the q multivariate levels, 
or a data frame/ matrix with two columns giving longnitude (first column), and latitude (second column).}

\item{kappa}{[\code{numeric}] Range parameter passed directly to \code{rSPDE::matern.covariance()}.}

\item{nu}{[\code{numeric}] Range parameter passed directly to \code{rSPDE::matern.covariance()}.}

\item{sigma_d2}{[\code{numeric}] Scaling variance parameter for the "matern block"}

\item{sigma_t2}{[\code{numeric}] Scaling variance parameter for the "trend/ season block"}

\item{trend}{[\code{logical}]. Set till "TRUE" if (one common) trend is to be included.}

\item{harmonic}{[\code{logical}]. Set till "TRUE" if (one common) seasonal effect is yto be included.}

\item{n_waves}{[\code{numeric}]. The number of harmonic waves.}

\item{regions}{[\code{factor}]. Vector giving different regions. Ignored if set to \code{NULL}}

\item{sigma}{[\code{numeric}] Range parameter passed directly to \code{rSPDE::matern.covariance()}.}
}
\value{
A block diagonal matrix whose upper left element has dimension [q x q], 
with structure \eqn{sigma_d^2} \eqn{C[nu](d)}, where \eqn{C[nu](d)} is the matern 
Covariance function and \eqn{sigma_d^2} is a positive scalar. The lower right
element is equal to  \eqn{sigma_t^2 I[trend == TRUE + 2 x n_waves]}.
}
\description{
The function \code{DLM_Get_W} constructs the W matrice. It is (by now) 
possible to include a multivariate response (dimention q), with matern covariance together 
 with common trend and common seasonality modeled as harmonic waves.
}
\details{
...
}
\examples{
\dontrun{

}


}
\author{
Lars Erik Gangsei\cr
lars.erik.gangsei@vetinst.no\cr
+47 950 61 231
}
