% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DLM_Multivariate_Normal.R
\name{DLM_Multivariate_Normal}
\alias{DLM_Multivariate_Normal}
\title{DLM_Multivariate_Normal: DLM for multivariate normal data.}
\usage{
DLM_Multivariate_Normal(
  Data,
  GG,
  FF,
  m0,
  C0,
  W = NULL,
  V,
  delta = 0.75,
  smoother_run = FALSE
)
}
\arguments{
\item{Data}{[\code{data.frame}]. Data frame with three elements:\cr
\code{t}: Vector with times/ dates.\cr
\code{y}: Matrix, where each row represents the observations for different 
levels at associated time step.}

\item{GG}{[\code{numeric matrix}] System matrix G. Assumed constant for all time elements.}

\item{FF}{[\code{numeric matrix}] Design matrix F. Assumed constant for all time elements.}

\item{m0}{[\code{numeric vector}] Prior mean for (latent) parameter vector \eqn{\theta}.}

\item{C0}{[\code{numeric matrix}] Prior variance for (latent) parameter vector \eqn{\theta}.}

\item{W}{[\code{numeric matrix}] System variance.}

\item{V}{[\code{numeric matrix}] Error variance.}

\item{delta}{[\code{numeric scalar}]. Discount factor. Only used if \code{W} is \code{NULL}.}

\item{smoother_run}{[\code{logical}]. If the DLM smoother is to be ran}
}
\value{
A list with elements at, Rt,mt,Ct,ft,Qt,pt,etaHat,vHat,Qst
}
\description{
The function \code{DGLM_Multivariate_Normal} fits a Dynamic 
Linear Model (DLM) to the data input.
}
\details{
...
}
\examples{
\dontrun{

}


}
\author{
Lars Erik Gangsei\cr
lars.erik.gangsei@vetinst.no\cr
+47 950 61 231
}
