% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DLM_simulate.R
\name{DLM_simulate}
\alias{DLM_simulate}
\title{DLM_simulate: DLM for multivariate normal data.}
\usage{
DLM_simulate(mu0, GG, FF, W, V, N = 100)
}
\arguments{
\item{mu0}{[\code{numeric vector}] Expected value for \eqn{\theta[0]}}

\item{GG}{[\code{numeric matrix}] System matrix G \code{size (q x q)}. 
Assumed constant for all time elements.}

\item{FF}{[\code{numeric matrix}] Design matrix F, \code{size (q x p)}. 
Assumed constant for all time elements.}

\item{W}{[\code{numeric matrix}] \code{size (q x q)}. System variance.}

\item{V}{[\code{numeric matrix}] \code{size (p x p)}. Error variance.}

\item{n}{[\code{integer}].}
}
\value{
A data frame with elements:\cr
}
\description{
The function \code{DGLM_simulate} simulate data from a Dynamic 
Linear Model (DLM).
}
\details{
...
}
\examples{
\dontrun{

}


}
\author{
Lars Erik Gangsei\cr
lars.erik.gangsei@vetinst.no\cr
+47 950 61 231
}
