% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NVI_Sea_distances.R
\name{NVI_Sea_distances}
\alias{NVI_Sea_distances}
\title{NVI_Sea_distances: Function to calculate Sea distances.}
\usage{
NVI_Sea_distances(
  sea_shape_path = file.path("//vetinst.no\\\\dfs-felles", "StasjonK", "FAG",
    "EksterneDatakilder", "Lokreg", "StotteData", "hav_300_100", "hdr.adf"),
  location,
  radius = 2e+05,
  resolution = 100
)
}
\arguments{
\item{sea_shape_path}{[\code{character}] path to file with shape-files
for the sea area.}

\item{location}{[\code{sf-object}?] containing one point. Needs to be same 
crs as used in \code{sea_shape_path}.}

\item{radius}{[\code{number}]. The radius for whithin distances should be 
calculated. In \code{meters}.}

\item{resolution}{[\code{number}]. Resolution/ size of raster pixels. 
In \code{meters}.}
}
\value{
To come....
}
\description{
The function \code{NVI_Sea_distances} returns a map with shortest 
distances from a given location to points locations in the sea.
}
\details{
Based on script from Lars Qviller, NVI
}
\examples{
\dontrun{

# Might come

}


}
\author{
Lars Erik Gangsei\cr
lars.erik.gangsei@vetinst.no\cr
+47 950 61 231
}
