% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScatterMap.R
\name{ScatterMap}
\alias{ScatterMap}
\title{ScatterMap: Points in a red - green scale on top of maps.}
\usage{
ScatterMap(
  Country_name = "Norway",
  Coord_lims = list(long = c(5, 35), lat = c(57, 72)),
  Plot_df,
  cex.point = 4,
  scale_lims = NULL
)
}
\arguments{
\item{Country_name}{[\code{character}] has to be listed in 
\code{rnaturalearth::ne_countries()$name}.}

\item{Coord_lims}{[\code{list}] with 2 elements.\code{lat} and \code{long}, 
both numeric vectors of length 2 with limits (in degrees) for latitude and 
longitude.}

\item{Plot_df}{[\code{list}].Data frame with the following elements:
\item \code{lat} [\code{numeric}]. Latitude for points to be plotted.
\item \code{long}[\code{numeric}].Longitude for points to be plotted.
\item \code{val}[\code{numeric}].Value for points to b illustrated.
\item \code{t}[\code{integer}].Indicator used to split figures with 
\code{ggplot2::facet_wrap()}.}

\item{cex.point}{[\code{numeric}]: point size. Passed to [\code{geom_point()}] 
as argument \code{size}.}

\item{scale_lims}{[\code{numeric}]: Vector of length two providing limits 
of the scale used in color grading}
}
\value{
To come....
}
\description{
The function \code{ScatterMap} returns a map (land/ ocean) with 
poinnts on top. Colors according to strength in input.
}
\details{
...
}
\examples{
\dontrun{

# Might come

}


}
\author{
Lars Erik Gangsei\cr
lars.erik.gangsei@vetinst.no\cr
+47 950 61 231
}
