% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TesnsorSplines.R
\name{TensorSplines}
\alias{TensorSplines}
\title{TensorSplines: Returns Spline matrix for for a 3D tensor product.}
\usage{
TensorSplines(
  XYZ,
  df = c(5, 5, 5),
  knots = NULL,
  degree = 3,
  intercept = TRUE,
  sparce = FALSE,
  dimnames = c("X", "Y", "Z")
)
}
\arguments{
\item{XYZ}{[\code{numeric matrix}] with 3 columns.}

\item{df}{[\code{integer}] vector with 3 elements.Passed to 
\code{splines2::bSpline()}, where first element is passed to first column of 
\code{XYZ} etc. Not used if \code{knots !=NULL}.}

\item{knots}{[\code{list}], with three elements. Each element passed to 
\code{splines2::bSpline()}, first element to first column of \code{XYZ} etc.
The elements of \code{knots} are numeric vectors. The first and last element
are used as argument \code{Boundary.knots} and the remaing elements as argument 
\code{knots} in \code{splines2::bSpline()}}

\item{degree}{[\code{integer}] passed directly to \code{splines2::bSpline()}}

\item{intercept}{[\code{logical}] passed directly to \code{splines2::bSpline()}}

\item{sparce}{[\code{logical}] set til \code{TRUE} if a sparce matrix is to be returned.}

\item{dimnames}{[code{character}] naming of the 'X', 'Y' and 'Z' variable}
}
\value{
A (possibly) sparce matrix returning the outer product defined by input, 
and evaluation of B(asic) splines as defined either by the knot sequenses and 
df (degrees of freedom).
}
\description{
The function \code{TensorSplines} returns a (sparse) matrix 
with tensor product for B-splines in 3 dimensions.
}
\details{
...
}
\examples{
\dontrun{

# Might come

}


}
\author{
Lars Erik Gangsei\cr
lars.erik.gangsei@vetinst.no\cr
+47 950 61 231
}
